Add a few small pieces to the puzzle:
• Install script - This should be run once after cloning on a raspbian
distro.  I've only tested this on a custom install that might already
have additional packages and the script might need more tweaking.

• Wanna-be-REST to UART implementation - Using the CherryPy library,
take PUT requests and send "a", "s", or "d" for directions to UART and
send "w" for whistle.

• Add an init script - This is run at boot to start our CherryPy python
webserver.  The server won't run properly unless the install script is
run which disables getty on ttyAMA0.