Send video at 320x240@30 fps with audio at 32 Kbps AAC to Twitch.TV.
I managed to get the audio functioning with video.  Ultimately the
culprit was the hardware h264 encoder.  For some reason the output is
incompatible with audio in a flash video container.   The work around is
to perform software encoding of both audio and video.  In order to avoid
buffer underruns (and program termination) is to lower video quality to
320x240.  So we have 320x240 video at 30 frames per second, with 32 Kbps
AAC audio.I managed to get the audio functioning with video.  Ultimately
the culprit was the hardware h264 encoder.  For some reason the output
is incompatible with audio in a flash video container.   The work around
is to perform software encoding of both audio and video.  In order to
avoid buffer underruns (and program termination) is to lower video
quality to 320x240.  So we have 320x240 video at 30 frames per second,
with 32 Kbps AAC audio.